Fix broken dependencies in template
I had mistakenly generated the requirements file without apache beam
which made all template runs fail. I have CI in here but it runs before
the template exists by executing the pipeline directly.

The mistake is fixed now and the CI is updated to upload a template
whenever the PR is submitted so that a test can be done using the
template.